Show all projects, instead of active ones.
The change respond to this task: https://github.com/iCarto/iom/issues/1

As you can see for the files involved, this is a project-wide action, as the
logic is spread all over the files with custom raw SQL queries like. A better
solution than dropping the predicate related to data would be to let the user
choose what to do: show all of active projects (maybe with a flag in the admin
backend).